Extract channel request show page components into separate modules (#4541)
Split the 1158-line show.ex into four focused files to improve
maintainability: helpers.ex (shared pure functions), components.ex
(reusable display components), timing.ex (timing visualization),
and show.ex (LiveView lifecycle + section composition).
